// client/src/app/interfaces/pageTypes.enum.ts
export enum PageTypes {
  TITLE = 'title',
  CODEONLY = 'codeOnly',
  TEXTONLY = 'textOnly',
  TEXTIMAGE = 'textImage',
  TEXTVIDEO = 'textVideo',
  VIDEOONLY = 'videoOnly',
}

export enum ImageOrVideoPosition {
  LEFT = 'left',
  RIGHT = 'right',
  BOTTOM = 'bottom',
  TOP = 'top',
  FULLSCREEN = 'fullscreen',
}

export enum ImageOrVideoSizes {
  HEIGHT = 'height',
  WIDTH = 'width',
  FULLSCREEN = 'fullscreen',
}